requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
requests==2.28.2
urllib3==1.26.15
yarg==0.1.9
discord.py==2.2.2
python-dotenv==0.21.0

